Move menu and pagination into its own component
